make sure there are non-x86 alternatives for the x86 asm!
